lab1/requirements.txt
pandas==1.4.2
plotly-express==0.4.1
scikit-learn==1.0.2
kaleido==0.2.1
